WIP: exe_wrapper: allow single exe_wrapper definition in cross, CLI and test_setup
Still WIP: many tests still fail.
I do not think its the right choice to try to convert all
`exe_wrapper` values to an ExternalProgram.
Additionally, this branch would also need more test coverage
and documentation to do with the exact behavior that meson uses
to combine multiple `exe_wrapper` arguments.
I chose to have all 3 `exe_wrapper` pathways be mutually exclusive
because that seemed like the simplest to implement, but
it might not be the right choice.